[#3] docker 연결해서 Match씬으로 가게 완료

+++ GameClient/Assets/02_Scripts/Inventory/InventoryManager.cs
@@ -83,7 +83,7 @@ public class InventoryManager : MonoBehaviour, IPacketReceiver
 
     public unsafe void OnPacketReceived(Packet packet)
     {
-        byte packetId = packet.pbase.packet_id;
+        ushort packetId = packet.pbase.packet_id;
 
         switch ((E_PACKET)packetId)
         {
